basic outline of view class: added method to created 2 players added method to choose one of the two players
