--- src/main/java/earthGame/testMain.java
@@ -1,7 +1,0 @@
-package earthGame;
-
-public class testMain {
-    public static void main(String[] args){
-        System.out.println("Hello World!");
-    }
-}
